Refactor services into domain-specific modules
