+++ src/services/admin/dashboard.ts
@@ -1,0 +1,5 @@
+import { apiClient } from '../client'
+
+export const adminDashboardService = {
+  getMetrics: () => apiClient.request('/api/admin/dashboard')
+}
