+++ src/services/analytics.ts
@@ -1,0 +1,5 @@
+import { apiClient } from './client'
+
+export const analyticsService = {
+  getMetrics: () => apiClient.request('/api/analytics/metrics')
+}
